requirements.txt
asgiref==3.9.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
daphne==4.2.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
huggingface-hub==0.35.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inquirerpy==0.3.4
intel-openmp==2021.4.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mkl==2021.4.0
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pfzy==0.3.4
pillow==11.3.0
prompt_toolkit==3.0.52
pyasn1_modules==0.4.2
pyasn1==0.6.1
pycparser==2.23
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tbb==2021.13.1
tokenizers==0.22.1
torch==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
transformers==4.56.2
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
zope.interface==8.0.1
